% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AitchisonKernel.R
\name{AitchisonKernel}
\alias{AitchisonKernel}
\title{Compute Aitchison Kernel}
\usage{
AitchisonKernel(Y)
}
\arguments{
\item{Y}{Matrix of microbiome count data with samples across rows and OTUs across columns.}
}
\value{
A list containing the kernel matrix, PCs, and PCA proportions.
}
\description{
This function computes the Aitchison kernel matrix from microbiome data.
}
