% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrayCurtisKernel.R
\name{BrayCurtisKernel}
\alias{BrayCurtisKernel}
\title{Compute Bray Curtis Kernel}
\usage{
BrayCurtisKernel(Y)
}
\arguments{
\item{Y}{Matrix of microbiome count data with samples across rows and OTUs across columns.}
}
\value{
A list containing the kernel matrix, PCs, and PCA proportions.
}
\description{
This function computes the Bray Curtis kernel matrix from microbiome data.
}
