% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plot.R
\name{create_plot}
\alias{create_plot}
\title{Create PCA Plot}
\usage{
create_plot(
  metadata,
  kernel_matrix,
  retain_prop = 0.9,
  covariates = NULL,
  breaks
)
}
\arguments{
\item{metadata}{Data frame containing the microbiome metadata. Must have columns \code{subjectid}, \code{time}, and other covariates.}

\item{kernel_matrix}{A kernel matrix derived from OTU data (e.g., Aitchison kernel).}

\item{retain_prop}{Proportion of variability explained by top l kernel PCs. Default is 0.9.}

\item{covariates}{A character vector of covariates to adjust for in the linear mixed model. Must be columns in \code{metadata}}

\item{breaks}{A single value indicating the number of time intervals to cut or a numeric vector with break points.}
}
\value{
A list containing a ggplot object, the full data frame with residuals, and the PCoA proportions.
}
\description{
This function creates a PCA plot of the standardized residuals based on metadata and a kernel matrix.
}
